Changed combine method to use a long inorder to return a Number.  This allows clients to get the required value from the Number object.
Fixed a bug where byte arrays containing a number which when left shifted caused a 1 to "land" on the 31st bit were negated in long form.
Updated classes using the combine method to reflect changes.

+++ hummingbird/space-system-model/src/main/java/com/logica/hummingbird/spacesystemmodel/parameters/behaviours/IntegerUnsignedBehaviour.java
@@ -34,7 +34,7 @@ public class IntegerUnsignedBehaviour extends AbstractIntegerBehaviour {
 		}
 		byte[] byteArray = BitSetUtility.toByteArray(actualParameter, this.SIZE_IN_BITS);
 		
-		long output = BytesUtility.combine(byteArray, this.SIZE_IN_BITS);
+		long output = BytesUtility.combine(byteArray, this.SIZE_IN_BITS).longValue();
 		LOG.debug("Testing combine.  Output(bin) = " + Long.toBinaryString(output));
 		LOG.debug("Testing combine.  Output(dec) = "  + output);
